aria/apps/visual-understanding/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aria-visual-understanding"
version = "1.0.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi==0.115.2",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "psycopg[binary]==3.2.3",
  "opencv-python-headless==4.10.0.84",
  "pillow==10.4.0",
  "numpy==2.1.2",
  "pytesseract==0.3.13",
  "httpx==0.27.2",
  "sqlalchemy==2.0.36",
  "redis==5.1.1",
  "opentelemetry-api==1.28.2",
  "opentelemetry-sdk==1.28.2",
  "opentelemetry-instrumentation-fastapi==0.49b2"
]
